requirements.txt
python-dotenv
telethon
dbt-postgres
psycopg2-binary
pandas

ultralytics
fastapi 
uvicorn 
sqlalchemy
pydantic
 